crate.version: Move `loadDepsTask` to the `version` model
