+++ app/templates/crate/version.hbs
@@ -192,7 +192,7 @@
             {{#each this.currentDependencies as |dep|}}
               <li><LinkToDep @dep={{dep}} /></li>
             {{else}}
-              {{#if this.loadDepsTask.isRunning}}
+              {{#if this.currentVersion.loadDepsTask.isRunning}}
                 <li>Loading…</li>
               {{else}}
                 <li>None</li>
